#46 Re-create signature metadata for PDF Signed Documents updating
Signature nodes are re-created for every updating
Signed aspect is removed if no signature is included in the new content
NOTICE: Due to this modification, signature metadata history is lost. It's not so relevant, as this signature metadata is extracted directly from the signature itself, so it can be recovered at any time from original file.